requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Vector database
qdrant-client>=1.7.0

# AI/ML
google-genai>=0.3.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Prisma client
prisma>=0.12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
httpx>=0.26.0
